+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
-## v0.1.1.9000
+## v0.1.3 (01/09/2022)
+
+### Fix
+
+- Fixed typing problems, mypy connected.
+- Fixed html documentation formatting.
+
+### Tests
+
+- Added more tests for dataclasses.
 
 ### Miscellaneous
 
+- Added more validation to dataclasses.
+- Improved documentation.
 - Removed `get_mappings` method from the `cat2cat_mappings` dataclass.
 - Removed not needed import calls.
 - Created the project description and sidebar on pypi.